DFR-2700 removing Org when not represented
