+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/noc/nocworkflows/UpdateRepresentationWorkflowService.java
@@ -6,7 +6,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.util.ObjectUtils;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseRole;
@@ -20,7 +19,6 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.service.SystemUserService;
 import java.util.Map;
 import java.util.Optional;
 
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.APPLICANT_FIRST_MIDDLE_NAME;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.APPLICANT_ORGANISATION_POLICY;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.APP_SOLICITOR_POLICY;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.CHANGE_ORGANISATION_REQUEST;
@@ -47,35 +45,24 @@ public class UpdateRepresentationWorkflowService {
                                                                              CaseDetails originalCaseDetails) {
         log.info("Received request to update representation on case with Case ID: {}", caseDetails.getId());
 
-        log.info("DEBUGGING NOC - handleNoticeOfChangeWorkflow entered and applicant name is still present {} with Case ID: {}",
-            ObjectUtils.nullSafeConciseToString(caseDetails.getData().get(APPLICANT_FIRST_MIDDLE_NAME)), caseDetails.getId());
-
         assignCaseAccessService.findAndRevokeCreatorRole(caseDetails);
 
-        log.info("DEBUGGING NOC - findAndRevokeCreatorRole entered and applicant name is still present {} with Case ID: {}",
-            ObjectUtils.nullSafeConciseToString(caseDetails.getData().get(APPLICANT_FIRST_MIDDLE_NAME)), caseDetails.getId());
         log.info("findAndRevokeCreatorRole executed {}", caseDetails.getId());
         Map<String, Object> caseData = noticeOfChangeService.updateRepresentation(caseDetails, authorisationToken,
             originalCaseDetails);
-        log.info("DEBUGGING NOC - noticeOfChangeService.updateRepresentation entered and applicant name is still present {} with Case ID: {}",
-            ObjectUtils.nullSafeConciseToString(caseDetails.getData().get(APPLICANT_FIRST_MIDDLE_NAME)), caseDetails.getId());
+
 
         if (isNoOrganisationsToAddOrRemove(caseDetails)) {
             persistDefaultOrganisationPolicy(caseDetails);
-            log.info("DEBUGGING NOC - persistDefaultOrganisationPolicy entered and applicant name is still present {} with Case ID: {}",
-                ObjectUtils.nullSafeConciseToString(caseDetails.getData().get(APPLICANT_FIRST_MIDDLE_NAME)), caseDetails.getId());
+
             setDefaultChangeOrganisationRequest(caseDetails);
-            log.info("DEBUGGING NOC - setDefaultChangeOrganisationRequest entered and applicant name is still present {} with Case ID: {}",
-                ObjectUtils.nullSafeConciseToString(caseDetails.getData().get(APPLICANT_FIRST_MIDDLE_NAME)), caseDetails.getId());
+
             return AboutToStartOrSubmitCallbackResponse.builder().data(caseData).build();
         }
 
         caseDetails.getData().putAll(caseData);
         caseDetails = noticeOfChangeService.persistOriginalOrgPoliciesWhenRevokingAccess(caseDetails,
             originalCaseDetails);
-        log.info("DEBUGGING NOC - noticeOfChangeService.persistOriginalOrgPoliciesWhenRevokingAccess "
-                + "entered and applicant name is still present {} with Case ID: {}",
-            ObjectUtils.nullSafeConciseToString(caseDetails.getData().get(APPLICANT_FIRST_MIDDLE_NAME)), caseDetails.getId());
 
         return assignCaseAccessService.applyDecision(systemUserService.getSysUserToken(), caseDetails);
     }
